Fixed Makefile to link linalg.c fixing other issues that came up in the process
